+++ app/build.gradle
@@ -75,4 +75,7 @@ dependencies {
     // Glide
     implementation 'com.github.bumptech.glide:glide:4.15.1'
     implementation 'jp.wasabeef:glide-transformations:4.3.0'
+
+    //카카오 로그인
+    implementation 'com.kakao.sdk:v2-user:2.10.0'
 }
