+++ app/src/main/AndroidManifest.xml
@@ -36,7 +36,7 @@
 
         </activity>
 
-        <!-- <activity
+        <activity
             android:name="com.kakao.sdk.auth.AuthCodeHandlerActivity"
             android:exported="true">
             <intent-filter>
@@ -49,7 +49,7 @@
                     android:host="oauth"
                     android:scheme="kakaoe9c2a8bf10ae12652fdc9ee9059ac02f" />
             </intent-filter>
-        </activity> -->
+        </activity>
 
     </application>
